Module data_encoding::base

Generic base module.

This module defines a generic interface, namely Base, and an optimized implementation, namely Opt, for positional numerical systems with radix ranging from 2 to 64 (powers of two only). Other base constraints are described in the Base interface.

Structs

Opt
Optimized implementation.

Spec
Specification implementation.

Enums

EqualError
Equality errors.

ValidError
Validity errors.

Traits

Base
Generic interface.

Functions

dec
Returns the decoding length of a base.

enc
Returns the encoding length of a base.

equal
Checks whether two bases are equal.

len
Returns the period length of a base.

mask
Returns the bit-mask of a base.

valid
Checks whether a base is valid.