Struct data_encoding::base::Spec

pub struct Spec {
    pub val: &'static [(u8, u8)],
    pub pad: u8,
}

Specification implementation.

This implementation uses an array of inclusive ranges to easily describe symbol to value assocation. It is not meant for performance but for specification using the equal function.

Fields
val: &'static [(u8, u8)]

Symbol to value association.

This array defines inclusive ranges of symbols in value order. These ranges must not overlap. For instance, for &[(b'0', b'9'), (b'A', b'F')], b'0' has value 0, b'8' has value 8, and b'E' has value 14.

pad: u8

The padding.